# go/api_response.go
/*
 * S3 asset uploader REST API
 *
 * an S3 asset uploader API
 *
 * API version: 1.0.0
 * Contact: bill.hongwu.chen@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type ApiResponse struct {

	Code int32 `json:"code,omitempty"`

	Type_ string `json:"type,omitempty"`

	Message string `json:"message,omitempty"`
}