# go/asset_status.go
/*
 * S3 asset uploader REST API
 *
 * an S3 asset uploader API
 *
 * API version: 1.0.0
 * Contact: bill.hongwu.chen@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type AssetStatus struct {

	// UUID of asset
	Id string `json:"id,omitempty"`

	// The status of file upload operation to s3
	Status string `json:"status,omitempty"`
}